requirements.txt
pyyaml==3.12
requests==2.18